cleanup information drawers
