+++ packages/desktop/views/dashboard/drawers/network-config/views/components/StardustInformationDrawer.svelte
@@ -1,9 +1,16 @@
 <script lang="ts">
     import { Button, IconName } from '@bloomwalletio/ui'
+    import { DrawerTemplate } from '@components'
     import { localize } from '@core/i18n'
+    import { Router } from '@core/router'
     import { PopupId, closePopup, openPopup } from '@desktop/auxiliary/popup'
     import ConfigureNodeList from './ConfigureNodeList.svelte'
     import LocalProofOfWork from './LocalProofOfWork.svelte'
+    import { IStardustNetwork } from '@core/network'
+    import { NetworkConfigRoute } from '../../network-config-route.enum'
+
+    export let drawerRouter: Router<NetworkConfigRoute>
+    export let network: IStardustNetwork
 
     let nodesContainer: HTMLElement
 
@@ -26,8 +33,17 @@
     }
 </script>
 
-<div class="flex flex-col w-full space-y-4">
-    <ConfigureNodeList bind:nodesContainer />
-    <LocalProofOfWork />
-</div>
-<Button variant="text" icon={IconName.Plus} width="full" text={localize('actions.addNode')} on:click={onAddNodeClick} />
+<DrawerTemplate title={network.name} {drawerRouter}>
+    <div class="flex flex-col w-full space-y-4 px-6">
+        <ConfigureNodeList bind:nodesContainer />
+        <LocalProofOfWork />
+    </div>
+    <Button
+        slot="footer"
+        variant="text"
+        icon={IconName.Plus}
+        width="full"
+        text={localize('actions.addNode')}
+        on:click={onAddNodeClick}
+    />
+</DrawerTemplate>
