+++ packages/desktop/views/dashboard/drawers/network-config/views/components/index.ts
@@ -1,3 +1,3 @@
 export { default as AddIscChainForm } from './AddIscChainForm.svelte'
-export { default as EvmChainInformation } from './EvmChainInformation.svelte'
-export { default as StardustInformation } from './StardustInformation.svelte'
+export { default as EvmChainInformationDrawer } from './EvmChainInformationDrawer.svelte'
+export { default as StardustInformationDrawer } from './StardustInformationDrawer.svelte'
